feature: added Weapon, Bullet, Satellite, Utility functions. Changed MainPlayer's x & y arguments to float

+++ src/main/java/it/unibo/geosurv/core/Block.java
@@ -15,12 +15,11 @@ public class Block extends GameObject {
 
     public void render(Graphics g) {
         g.setColor(Color.black);
-        g.fillRect(x, y, 32, 32);
+        g.fillRect((int)x, (int)y, 32, 32);
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 32, 32);
-        
+        return new Rectangle((int)x, (int)y, 32, 32);
     }
     
 }
